Move check_cclib to common file
